backend/requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
anyio==4.10.0
appnope==0.1.4
asgiref==3.9.1
asttokens==3.0.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.3.0
autobahn==24.4.2
automat==25.4.16
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
channels==4.3.1
channels-redis==4.3.0
charset-normalizer==3.4.3
comm==0.2.3
constantly==23.10.4
cryptography==44.0.3
daphne==4.2.1
debugpy==1.8.16
decorator==5.2.1
distro==1.9.0
django==5.2.5
django-cors-headers==4.7.0
djangorestframework==3.16.1
executing==2.2.0
filelock==3.19.1
flatbuffers==25.2.10
fsspec==2025.7.0
gast==0.4.0
google-auth==2.40.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
groq==0.31.0
grpcio==1.74.0
h11==0.16.0
h5py==3.14.0
hf-xet==1.1.8
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
ipykernel==6.30.1
ipython==8.37.0
ipython-pygments-lexers==1.1.1
jedi==0.19.2
jiter==0.10.0
jupyter-client==8.6.3
jupyter-core==5.8.1
keras==2.13.1
libclang==18.1.1
markdown==3.8.2
markupsafe==3.0.2
matplotlib-inline==0.1.7
msgpack==1.1.1
nest-asyncio==1.6.0
numpy==1.24.3
oauthlib==3.3.1
openai==1.101.0
opt-einsum==3.4.0
packaging==25.0
parso==0.8.5
pexpect==4.9.0
platformdirs==4.4.0
prompt-toolkit==3.0.51
protobuf==4.25.8
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pydub==0.25.1
pygments==2.19.2
pyopenssl==24.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pyyaml==6.0.2
pyzmq==27.0.2
redis==6.4.0
regex==2025.7.34
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
safetensors==0.6.2
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
speechrecognition==3.14.3
sqlparse==0.5.3
stack-data==0.6.3
termcolor==3.1.0
tokenizers==0.21.4
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
twisted==25.5.0
txaio==25.6.1
typing-extensions==4.15.0
typing-inspection==0.4.1
urllib3==2.5.0
wcwidth==0.2.13
werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.3
zope-interface==7.2
pyjwt==2.8.0
djangorestframework-simplejwt==5.3.0
